requirements.txt
openai==1.35.3
azure-identity==1.16.0
azure-mgmt-resource==23.0.0
azure-mgmt-network==25.1.0
PyGithub==2.3.0
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0
jinja2==3.1.4
mcp
azure-identity
openai
logging